requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# AI/ML - OpenAI API (fast and lightweight)
openai>=1.30.0

# News Data & Processing
aiohttp==3.10.11
feedparser==6.0.11

# ML for Bias Analysis & Clustering
scikit-learn==1.5.2
numpy==2.3.2
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1

# Utilities
python-dotenv==1.1.1
structlog==25.4.0